Extract a shared GameListView from the three games-list surfaces
The games page, a user's match history, and a league's games tab were three
near-verbatim copies of the same body: the filter-bar wiring, the ~50-line
keyboard-nav useKeyListener, selectIndex/moveSelection, the useGameListSearch
plumbing, and the confirmed-empty/list/side-panel render. They had already
drifted (the games page's setFormat didn't clear the matchup param the way the
other two did).

Pull that body into client/games/game-list-view.tsx. Each surface is now a thin
wrapper that supplies a loadPage (dispatching getGames / getMatchHistory /
getLeagueGames) plus a few presentation flags (showRankedCustom, showResult,
forUserId) and its own outer container; GameListView owns the filter state
(mirrored to URL params), selection, and keyboard handling. Keyboard nav,
empty-state, and filter behavior now live in one place.

Verified end-to-end in the app against real data: all three surfaces render,
filter (ranked/custom shown only on match history), select into the detail
panel, and show win/loss only on match history.
