+++ client/games/game-list-view.tsx
@@ -1,0 +1,372 @@
+import { useRef, useState } from 'react'
+import { useTranslation } from 'react-i18next'
+import styled from 'styled-components'
+import {
+  EncodedMatchupString,
+  GameDurationFilter,
+  GameFormat,
+  GameSortOption,
+} from '../../common/games/game-filters'
+import { SbUserId } from '../../common/users/sb-user-id'
+import { useContextMenu } from '../dom/use-context-menu'
+import { useKeyListener } from '../keyboard/key-listener'
+import InfiniteScrollList from '../lists/infinite-scroll-list'
+import { Popover } from '../material/popover'
+import { useLocationSearchParam } from '../navigation/router-hooks'
+import { useUserLocalStorageValue } from '../react/state-hooks'
+import { bodyLarge } from '../styles/typography'
+import { navigateToGameResults } from './action-creators'
+import { renderGamesWithDayHeaders, resolveDateRangeMs } from './day-header'
+import { GameContextMenuContent } from './game-context-menu'
+import { GameFilterBar } from './game-filter-bar'
+import { isDateSort, parseDuration, parseFormat, parseMatchup, parseSort } from './game-filter-url'
+import { GameListEntry } from './game-list-entry'
+import { GameRecordSidePanel } from './game-record-side-panel'
+import { GameListSearchPage, useGameListSearch } from './use-game-list-search'
+
+const NoResults = styled.div`
+  ${bodyLarge};
+
+  color: var(--theme-on-surface-variant);
+`
+
+const ErrorText = styled.div`
+  ${bodyLarge};
+
+  color: var(--theme-error);
+`
+
+const BodyRow = styled.div`
+  display: flex;
+  flex-direction: row;
+  gap: 24px;
+`
+
+const ListColumn = styled.div`
+  flex-grow: 1;
+  min-width: 0;
+`
+
+/** The normalized, ready-to-request filter values a game list view is currently showing. */
+export interface GameListFilters {
+  ranked?: boolean
+  custom?: boolean
+  duration?: GameDurationFilter
+  sort?: GameSortOption
+  mapName?: string
+  playerName?: string
+  format?: GameFormat
+  matchup?: EncodedMatchupString
+  /** Inclusive lower bound (unix ms) on the game's start time. */
+  startDate?: number
+  /** Inclusive upper bound (unix ms) on the game's start time. */
+  endDate?: number
+}
+
+export interface GameListViewProps {
+  /**
+   * Loads one page of games for the given filters/offset — implemented per surface (the games
+   * page, a user's match history, a league's games) by dispatching the matching request. Resolves
+   * with the page's game ids and whether more remain; populating the normalized game/map/user
+   * stores is a side effect of that dispatch.
+   */
+  loadPage: (
+    filters: GameListFilters,
+    offset: number,
+    signal: AbortSignal,
+  ) => Promise<GameListSearchPage>
+  /** Shows the Ranked/Custom source toggles and reads their URL params (match history only). */
+  showRankedCustom?: boolean
+  /** Shows each row's win/loss result, from `forUserId`'s perspective (match history only). */
+  showResult?: boolean
+  /** Whose perspective results and the side panel roster are shown from (match history only). */
+  forUserId?: SbUserId
+  /** Rendered in place of the list once a load has confirmed there are no matching games. */
+  noResultsText: string
+  /** Rendered in place of the list when a load fails. */
+  errorText: string
+}
+
+/**
+ * The shared games-list surface: a filter bar over a paginated, selectable list with a detail side
+ * panel and per-row context menu, backed by `useGameListSearch`. The games page, match history, and
+ * a league's games tab are all thin wrappers that supply a `loadPage` and their own outer layout.
+ *
+ * Owns the filter state (mirrored to URL search params), the list selection, and the keyboard
+ * navigation; callers differ only in where they fetch from and a few presentation flags. Renders as
+ * a fragment (filter bar, then the list/panel row) so each surface controls its own container and
+ * spacing.
+ */
+export function GameListView({
+  loadPage,
+  showRankedCustom = false,
+  showResult = false,
+  forUserId,
+  noResultsText,
+  errorText,
+}: GameListViewProps) {
+  const { t } = useTranslation()
+
+  const [rankedParam, setRankedParam] = useLocationSearchParam('ranked')
+  const [customParam, setCustomParam] = useLocationSearchParam('custom')
+  const [durationParam, setDurationParam] = useLocationSearchParam('duration')
+  const [sortParam, setSortParam] = useLocationSearchParam('sort')
+  const [mapName, setMapNameParam] = useLocationSearchParam('mapName')
+  const [playerName, setPlayerNameParam] = useLocationSearchParam('playerName')
+  const [formatParam, setFormatParam] = useLocationSearchParam('format')
+  const [matchupParam, setMatchupParam] = useLocationSearchParam('matchup')
+  const [startDateParam, setStartDateParam] = useLocationSearchParam('startDate')
+  const [endDateParam, setEndDateParam] = useLocationSearchParam('endDate')
+
+  // Ranked/custom only exist on the match-history surface; elsewhere we neither read nor send them,
+  // so a hand-edited `?ranked=true` on another surface can't leak into the request.
+  const ranked = showRankedCustom && rankedParam === 'true'
+  const custom = showRankedCustom && customParam === 'true'
+  const duration = parseDuration(durationParam)
+  const sort = parseSort(sortParam)
+  const format = parseFormat(formatParam)
+  const matchup = parseMatchup(matchupParam, format)
+
+  const [selectedId, setSelectedId] = useState<string>()
+  const rowElemsRef = useRef(new Map<string, HTMLDivElement>())
+
+  // Remembered per-user and shared across the replay library, games page, match history, and league
+  // games: hides the game length and (where shown) the match result — both spoilers — from list rows.
+  const [spoilerFree, setSpoilerFree] = useUserLocalStorageValue('gamesSpoilerFree', false)
+
+  const { onContextMenu, contextMenuPopoverProps } = useContextMenu()
+
+  const loadPageForSearch = (offset: number, signal: AbortSignal): Promise<GameListSearchPage> => {
+    const { startMs, endMs } = resolveDateRangeMs(startDateParam, endDateParam)
+    return loadPage(
+      {
+        ranked: ranked || undefined,
+        custom: custom || undefined,
+        duration: duration === GameDurationFilter.All ? undefined : duration,
+        sort: sort === GameSortOption.LatestFirst ? undefined : sort,
+        mapName: mapName || undefined,
+        playerName: playerName || undefined,
+        format,
+        matchup,
+        startDate: startMs,
+        endDate: endMs,
+      },
+      offset,
+      signal,
+    )
+  }
+
+  const { games, hasMoreGames, isLoadingMore, searchError, refreshToken, reset, onLoadMore } =
+    useGameListSearch(loadPageForSearch)
+
+  const selectedGame = games.find(g => g.id === selectedId) ?? games[0]
+  const selectedIndex = selectedGame ? games.findIndex(g => g.id === selectedGame.id) : -1
+  const sortIsDateBased = isDateSort(sort)
+
+  const selectIndex = (index: number) => {
+    if (index < 0 || index >= games.length) return
+    const game = games[index]
+    setSelectedId(game.id)
+    rowElemsRef.current.get(game.id)?.scrollIntoView({ block: 'nearest' })
+  }
+  const moveSelection = (delta: number) => {
+    if (games.length === 0) return
+    const base = selectedIndex < 0 ? 0 : selectedIndex
+    const next = Math.min(Math.max(base + delta, 0), games.length - 1)
+    selectIndex(next)
+  }
+
+  useKeyListener({
+    onKeyDown: (event: KeyboardEvent) => {
+      // Every key here acts on the games list, but this page's key listener boundary is shared with
+      // other UI (e.g. the social sidebar's chat input, whose keydowns bubble to the document).
+      // While a text-entry element has focus, the keystroke belongs to it instead.
+      const target = event.target
+      if (
+        target instanceof HTMLElement &&
+        (target.tagName === 'INPUT' ||
+          target.tagName === 'TEXTAREA' ||
+          target.tagName === 'SELECT' ||
+          target.isContentEditable)
+      ) {
+        return false
+      }
+
+      switch (event.code) {
+        case 'ArrowUp':
+          moveSelection(-1)
+          return true
+        case 'ArrowDown':
+          moveSelection(1)
+          return true
+        case 'PageUp':
+          moveSelection(-10)
+          return true
+        case 'PageDown':
+          moveSelection(10)
+          return true
+        case 'Home':
+          selectIndex(0)
+          return true
+        case 'End':
+          selectIndex(games.length - 1)
+          return true
+        case 'Enter':
+        case 'NumpadEnter': {
+          const active = document.activeElement
+          // If the user is on an interactive control (e.g. a focused button), let it handle Enter.
+          if (
+            active instanceof HTMLElement &&
+            (active.tagName === 'BUTTON' || active.tagName === 'A')
+          ) {
+            return false
+          }
+          if (selectedGame) navigateToGameResults(selectedGame.id)
+          return true
+        }
+      }
+
+      return false
+    },
+  })
+
+  const filterBar = (
+    <GameFilterBar
+      showRankedCustom={showRankedCustom}
+      ranked={ranked}
+      setRanked={v => {
+        setRankedParam(v ? 'true' : '')
+        reset()
+      }}
+      custom={custom}
+      setCustom={v => {
+        setCustomParam(v ? 'true' : '')
+        reset()
+      }}
+      duration={duration}
+      setDuration={v => {
+        setDurationParam(v === GameDurationFilter.All ? '' : v)
+        reset()
+      }}
+      sort={sort}
+      setSort={v => {
+        setSortParam(v === GameSortOption.LatestFirst ? '' : v)
+        reset()
+      }}
+      mapName={mapName}
+      setMapName={v => {
+        setMapNameParam(v)
+        reset()
+      }}
+      playerName={playerName}
+      setPlayerName={v => {
+        setPlayerNameParam(v)
+        reset()
+      }}
+      format={format}
+      setFormat={v => {
+        setFormatParam(v ?? '')
+        // A matchup is tied to a specific format (team size), so any existing value no longer
+        // applies once the format changes. Clear it so it doesn't linger in the URL as cruft.
+        setMatchupParam('')
+        reset()
+      }}
+      matchup={matchup}
+      setMatchup={v => {
+        setMatchupParam(v ?? '')
+        reset()
+      }}
+      spoilerFree={spoilerFree}
+      setSpoilerFree={setSpoilerFree}
+      startDate={startDateParam}
+      setStartDate={v => {
+        setStartDateParam(v)
+        reset()
+      }}
+      endDate={endDateParam}
+      setEndDate={v => {
+        setEndDateParam(v)
+        reset()
+      }}
+    />
+  )
+
+  // A page load that's returned at least once with nothing left to fetch is a confirmed empty
+  // result; until then (including the very first, still in-flight page) we render the list shell so
+  // its own loading indicator can show.
+  const confirmedEmpty = !hasMoreGames && games.length === 0
+
+  let listBody: React.ReactNode
+  if (searchError) {
+    listBody = <ErrorText>{errorText}</ErrorText>
+  } else if (confirmedEmpty) {
+    listBody = <NoResults>{noResultsText}</NoResults>
+  } else {
+    const gameItems = renderGamesWithDayHeaders(games, sort, t, game => (
+      <GameListEntry
+        key={game.id}
+        game={game}
+        showResult={showResult}
+        forUserId={forUserId}
+        spoilerFree={spoilerFree}
+        selected={game.id === selectedGame?.id}
+        onClick={setSelectedId}
+        onDoubleClick={gameId => navigateToGameResults(gameId)}
+        onContextMenu={(gameId, event) => {
+          setSelectedId(gameId)
+          onContextMenu(event)
+        }}
+        ref={el => {
+          if (el) {
+            rowElemsRef.current.set(game.id, el)
+          } else {
+            rowElemsRef.current.delete(game.id)
+          }
+        }}
+      />
+    ))
+
+    listBody = (
+      <InfiniteScrollList
+        nextLoadingEnabled={true}
+        isLoadingNext={isLoadingMore}
+        hasNextData={hasMoreGames}
+        refreshToken={refreshToken}
+        onLoadNextData={onLoadMore}>
+        {gameItems}
+      </InfiniteScrollList>
+    )
+  }
+
+  // Mirrors the replay library's inspector: dropped entirely once there are confirmed to be no
+  // results, rather than showing an empty "select a game" placeholder beside the empty-state message.
+  const showPanel = !confirmedEmpty && !searchError
+
+  return (
+    <>
+      {filterBar}
+
+      <BodyRow>
+        <ListColumn>{listBody}</ListColumn>
+
+        {showPanel ? (
+          <GameRecordSidePanel
+            game={selectedGame}
+            forUserId={forUserId}
+            alignWithFirstRow={sortIsDateBased}
+            onViewResults={gameId => navigateToGameResults(gameId)}
+          />
+        ) : null}
+      </BodyRow>
+
+      {selectedGame ? (
+        <Popover {...contextMenuPopoverProps}>
+          <GameContextMenuContent
+            game={selectedGame}
+            onDismiss={contextMenuPopoverProps.onDismiss}
+          />
+        </Popover>
+      ) : null}
+    </>
+  )
+}
